Never select widget suggestions for Copilot compat
GitHub Copilot suggestions (the suggestions shown in ghost text when using
Copilot)  are accepted with Tab. Tab also also accepts the selected suggestion
in the suggestion widget 😫, blocking the single-Tab acceptance of Copilot
suggestions.

This change prevents the selection of the first suggestion in the suggestion
widget, which makes it again possible to accept the Copilot ghost text
suggestion with a single press of the Tab key.