Revert update to VS Code group size redistribution
VS Code June 2023 included a change to the default setting of
workbench.editor.splitSizing which caused new groups to not redistribute the
size of resized editor groups:

https://code.visualstudio.com/updates/v1_80#_editor-group-split-sizing-changed-to-auto

This means that if a user performs any resizing of editor groups (the tab panes
in the editor) - which I do a lot - then switching to a different layout by
opening a new editor group (eg. dragging a file to the edge of an editor group
to open a new editor group) does not redistribute the space between all of the
tab groups. The size redistribution is a feature I was relying upon, so this
changes the setting back to restore the previous behavior.